requirements.txt
numpy==1.23.5 
scikit-learn==1.2.2
apache-airflow-providers-common-sql>=1.20.0
psycopg2-binary>=2.9.4

